pyproject.toml
[tool.poetry]
name = "home-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.11.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '\.git|\.venv|\.idea|__pycache__|build|dist'

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'
